cli: split subcommands logic out of main
This splits and moves subcommands logic out of main, into their own
`cli` modules.
